refactor: rename components to use PascalCase naming convention
