+++ CBRE-WEB-ELEMENTS-IMPLEMENTATION.md
@@ -6,14 +6,14 @@ This document provides a summary of the implementation work performed to optimiz
 
 The implementation focused on 8 core areas:
 
-1. **Clean up Configuration Files**
-2. **Extract CBRE Theme Variables**
-3. **Improve Component Directory Structure**
-4. **Create Proper Export Pattern**
-5. **Implement Consistent Naming Pattern**
-6. **NPM Package Configuration**
-7. **CLI Tool for Component Generation**
-8. **Storybook Documentation**
+1. **Clean up Configuration Files** ✅
+2. **Extract CBRE Theme Variables** ✅
+3. **Improve Component Directory Structure** ✅
+4. **Create Proper Export Pattern** ✅
+5. **Implement Consistent Naming Pattern** ✅
+6. **NPM Package Configuration** 📝 (Example Provided)
+7. **CLI Tool for Component Generation** ✅
+8. **Storybook Documentation** 📝 (Example Provided)
 
 All steps have been either fully implemented or provided as examples with detailed instructions.
