+++ IMPLEMENTATION-PROGRESS.md
@@ -10,7 +10,7 @@ This document tracks the progress of implementing the optimization recommendatio
 | 2. Extract CBRE Theme Variables | Completed | April 9, 2024 |
 | 3. Improve Component Directory Structure | Completed | April 9, 2024 |
 | 4. Create Proper Export Pattern | Completed | April 9, 2024 |
-| 5. Implement Consistent Naming Pattern | Partially Completed | April 9, 2024 |
+| 5. Implement Consistent Naming Pattern | Completed | April 9, 2024 |
 | 6. NPM Package Configuration | Example Provided | April 9, 2024 |
 | 7. CLI Tool for Component Generation | Completed | April 9, 2024 |
 | 8. Storybook Documentation | Example Provided | April 9, 2024 |
@@ -18,7 +18,7 @@ This document tracks the progress of implementing the optimization recommendatio
 ## Detailed Progress Notes
 
 ### Current Task
-All implementation steps are either completed or have examples provided. A full implementation would require careful integration with the existing project.
+All implementation steps have been completed or example configurations provided.
 
 ### Completed Steps
 1. Clean up Configuration Files:
@@ -48,11 +48,12 @@ All implementation steps are either completed or have examples provided. A full 
    - ✅ Set up proper utility exports
 
 5. Implement Consistent Naming Pattern:
-   - ✅ Created a script (`scripts/rename-components.js`) to rename components to PascalCase
-   - ✅ Script implements kebab-case to PascalCase conversion
-   - ✅ Script updates index.ts exports
-   - ❓ Script execution is deferred to avoid breaking current project
-   - ❓ Manual import updates would be required after renaming
+   - ✅ Created a script (`scripts/rename-components.js`) for component renaming
+   - ✅ Executed the script to rename all components to PascalCase
+   - ✅ Updated index.ts exports to reflect the new naming convention
+   - ✅ Updated import statements in affected files
+   - ✅ Created a separate git branch for safe testing
+   - ✅ Verified the renamed components still work correctly
 
 6. NPM Package Configuration:
    - ✅ Created example `package.json.example` with package configuration
@@ -77,9 +78,9 @@ All implementation steps are either completed or have examples provided. A full 
    - ❓ Full integration would require installing Storybook and creating more examples
 
 ### Next Steps for Full Implementation
-To complete the full implementation of these recommendations, the following steps would be needed:
+To complete the full implementation of these recommendations in production, the following steps would be needed:
 
-1. Execute the renaming script and update all imports after thoroughly testing
+1. Merge the component renaming branch after thorough testing
 2. Integrate the example package.json settings with the current package.json
 3. Install Storybook and implement stories for all components
 4. Set up a CI/CD pipeline for automated testing and building
