+++ src/components/blocks/index.ts
@@ -3,5 +3,5 @@
  * This file exports higher-level block components that compose multiple UI elements
  */
 
-export * from './cbre-cta-block';
-export * from './cbre-quote-block'; 
+export * from './CBRECtaBlock';
+export * from './CBREQuoteBlock';
