+++ src/components/cbre/CBREResizable.tsx
