+++ src/components/cbre/index.ts
@@ -3,26 +3,26 @@
  * This file exports all CBRE-specific components to simplify imports
  */
 
-export * from './cbre-accordion';
-export * from './cbre-arrow-button';
-export * from './cbre-badge';
-export * from './cbre-button';
-export * from './cbre-card';
-export * from './cbre-chart';
-export * from './cbre-checkbox';
-export * from './cbre-data-table';
-export * from './cbre-date-picker';
-export * from './cbre-dropdown-menu';
-export * from './cbre-hover-card';
-export * from './cbre-resizable';
-export * from './cbre-select';
-export * from './cbre-separator';
-export * from './cbre-sidebar';
-export * from './cbre-styled-card';
-export * from './cbre-table';
-export * from './cbre-tabs';
-export * from './cbre-theme-provider';
-export * from './cbre-toast';
-export * from './cbre-toggle';
-export * from './cbre-toggle-group';
-export * from './cbre-tooltip'; 
+export * from './CBREAccordion';
+export * from './CBREArrowButton';
+export * from './CBREBadge';
+export * from './CBREButton';
+export * from './CBRECard';
+export * from './CBREChart';
+export * from './CBRECheckbox';
+export * from './CBREDataTable';
+export * from './CBREDatePicker';
+export * from './CBREDropdownMenu';
+export * from './CBREHoverCard';
+export * from './CBREResizable';
+export * from './CBRESelect';
+export * from './CBRESeparator';
+export * from './CBRESidebar';
+export * from './CBREStyledCard';
+export * from './CBRETable';
+export * from './CBRETabs';
+export * from './CBREThemeProvider';
+export * from './CBREToast';
+export * from './CBREToggle';
+export * from './CBREToggleGroup';
+export * from './CBRETooltip';
